Fixing property controller methods, (get, update, destroy)

+++ src/api/controllers/property.js
@@ -1,4 +1,4 @@
-const { Property, Canton, MeterService } = require("../models/");
+const { Property, Canton, MeterService, sequelize } = require("../models/");
 const {
   errorResponse,
   successResponse,
@@ -21,6 +21,13 @@ const getAll = async (req, res) => {
   try {
     const properties = await Property.findAll({
       where: { idOwner: req.user?.id },
+      attributes: {
+        exclude: ["idOwner", "idMeterServices"],
+      },
+      include: {
+        model: MeterService,
+        as: "meterServices",
+      },
     });
 
     return res
@@ -50,28 +57,31 @@ const create = async (req, res) => {
         error.details.map((e) => e.message)
       )
     );
-
   try {
-    const { idCanton, lightMeter, waterMeter } = req.body;
-    const canton = await Canton.findOne({ where: { id: idCanton } });
+    const canton = await Canton.findOne({ where: { id: req.body.idCanton } });
 
     if (!canton)
       return res
         .status(NOT_FOUND)
         .json(errorResponse(res.statusCode, "Canton does not exists!"));
 
-    const meterServices =
-      lightMeter || waterMeter
-        ? await MeterService.create({
-            lightMeter: lightMeter,
-            waterMeter: waterMeter,
-          })
-        : null;
+    const newProperty = await sequelize.transaction(async (t) => {
+      const meterServices = await MeterService.create(
+        {
+          lightMeter: req.body?.lightMeter,
+          waterMeter: req.body?.waterMeter,
+        },
+        { transaction: t }
+      );
 
-    const newProperty = await Property.create({
-      ...req.body,
-      idOwner: req.user?.id,
-      idMeterServices: meterServices?.id,
+      return Property.create(
+        {
+          ...req.body,
+          idOwner: req.user?.id,
+          idMeterServices: meterServices.id,
+        },
+        { transaction: t }
+      );
     });
 
     return res
@@ -84,7 +94,6 @@ const create = async (req, res) => {
         )
       );
   } catch (e) {
-    console.log(e.message);
     return res
       .status(INTERNAL_SERVER_ERROR)
       .json(errorResponse(res.statusCode, "Server error"));
@@ -102,23 +111,51 @@ const update = async (req, res) => {
       )
     );
 
+  const t = await sequelize.transaction();
+
   try {
     const id = req.params.id;
 
-    const [num] = await Property.update(req.body, {
-      where: { id, idOwner: req?.user.id },
+    let property = await Property.findOne({
+      where: { id, idOwner: req.user?.id },
     });
 
-    if (num != 1)
+    if (!property)
       return res
         .status(NOT_FOUND)
-        .json(errorResponse(res.statusCode, "Cannot update property!"));
+        .json(errorResponse(res.statusCode, "Cannot find property!"));
 
+    property = Object.assign(property, { ...req.body });
+
+    const updatedProperty = await property.save({ transaction: t });
+
+    if (req.body.lightMeter || req.body.waterMeter) {
+      const [numMeterServices] = await MeterService.update(
+        {
+          lightMeter: req.body.lightMeter,
+          waterMeter: req.body.waterMeter,
+        },
+        {
+          where: {
+            id: updatedProperty?.idMeterServices,
+          },
+          transaction: t,
+        }
+      );
+
+      if (numMeterServices != 1) {
+        await t.rollback();
+        return res
+          .status(NOT_FOUND)
+          .json(errorResponse(res.statusCode, "Cannot update property!"));
+      }
+    }
+    await t.commit();
     return res
       .status(OK)
-      .json(successResponse(res.statusCode, "Property updated successfully!"));
+      .json(successResponse(res.statusCode, "Property updated!"));
   } catch (e) {
-    console.log(e.message);
+    await t.rollback();
     return res
       .status(INTERNAL_SERVER_ERROR)
       .json(errorResponse(res.statusCode, "Server error"));
@@ -126,22 +163,46 @@ const update = async (req, res) => {
 };
 
 const destroy = async (req, res) => {
+  const t = await sequelize.transaction();
+
   try {
     const id = req.params.id;
-    const num = await Property.destroy({
-      where: { id, idOwner: req?.user.id },
+    let property = await Property.findOne({
+      where: { id, idOwner: req.user?.id },
     });
 
-    if (num != 1)
+    if (!property)
+      return res
+        .status(NOT_FOUND)
+        .json(errorResponse(res.statusCode, "Property not found!"));
+
+    const numProperty = await Property.destroy({
+      where: { id, idOwner: req?.user.id },
+      transaction: t,
+    });
+
+    const numMeterService = await MeterService.destroy({
+      where: {
+        id: property?.idMeterServices,
+      },
+      transaction: t,
+    });
+
+    if (numProperty != 1 || numMeterService != 1) {
+      await t.rollback();
       return res
         .status(NOT_FOUND)
         .json(errorResponse(res.statusCode, "Cannot delete property!"));
+    }
+
+    await t.commit();
 
     return res
       .status(OK)
       .json(successResponse(res.statusCode, "Property deleted successfully!"));
   } catch (err) {
     console.log(e.message);
+    await t.rollback();
     return res
       .status(INTERNAL_SERVER_ERROR)
       .json(errorResponse(res.statusCode, "Server error"));
